Move Batcher under SessionKeeper
Batcher was acting as a separate component however
it was only used by SessionKeeper, this made it to be
in this weird place where it's isnt an active component
and also it's just a SessionKeeper just with thresholds.

To unify the code a bit, batcher is moved under telio-traversal
nearby SessionKeeper.

What's more BatcherTrait is introduced  that allows for mocked
expectations of SessionKeeper actually calling it.

Notifier for task add is also removed since wg_consolidate
call will force the wait_with_update to be recalled again,
so the batcher is simplified even further and adheres to libtelio
better.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>