Bump libtelio build to v2.6.2
test_cleanup.py was changed, due to removal of
additional controller from windows VM, which was
accidenatlly introduced by disabling qemu-guest-agent
in previous version (causing libvirt to automatically
add one more controller for host and guest communication).

So initally it was 5 devices, with v2.6.1 another
device appeared, increasing number to 6,
but with v2.6.2 it is again removed, returning back to 5 devices.