Move out SessionKeeper from Direct struct
SessionKeeper is solely used for Direct connections
and thus only enabled when Direct feature is enabled.

However since it is very minimal, it makes sense
to augment it for use in batching more things.

This commit moves out SessionKeeper out from Direct
struct to the parent structure and is always present
from now on.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>