Refactor and unit test SessionKeeper more
SessionKeeper log refactored to be easier to
follow. Also introduces calls to SessionKeeper
if batching is enabled when upgrading from
Proxying to Direct and vice-versa. Current solution
because the key is the same skips that and misaligns
the keepalives which is an easy win.

A bunch of testcases are augmented with batching parameter

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>