Enable coredumps when killing stuck TCPdump processes
In NatLab we observed that sometimes TCPdump might get stuck
when starting(waiting for 'listening on' output). In order
to peek into what might be happening, coredumps are enabled
and the /var/crash directory is mapped to the container.

NatLab runner code already collects coredumps from /var/crash so
they should just appear in the job artifacts.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>