Use event waiting instead of sleeping in test_event_content_meshnet
Sleep are generally bad when we're waiting for stuff since
the event might happen quicker but we still sleep thus
it's slower test.

They also are a source of flakyness, for example
sleeping for 5seconds and missing an event which happens at t=6
would fail the test in a flaky manner since it still happened
but after when we expected. Meanwhile an event trigger would succeed
if the timeout is big enough. In sleep case once flakyness
is observed, we tend to increase the sleep to avoid flakyness
but that in turns makes testcases longer to execute.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>