Wait for interface to report Connected state
Sometimes, windows IP set commands via netsh
don't actually set an IP even though command
does not report non-zero exit code and
both stdout and stderr are empty.

Experimentation revealed that the interface may
not yet be in a `Connected` state at the time
of the IP set action, so we enforce that before
issuing IP set commands by waiting for the state
to be reported.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>