Fix PQ restart logic for non-NepTUN adapters
NepTUN reports time_since_last_handshake as None if more than 180s
passed whereas other implementations keep reporting it as an actual
time. A timestamp was recently added to the PQ entity to fix a race
condition in the current logic, but that timestamp can be used as the
only deciding factor on whether to restart PQ, so this commit starts
doing exactly that. It also parameterizes the regression test to make
sure it works on all adapter types