Add granular timestamps to logs
There is a suspicion that our logs can lock from time to time.
Currently the timestamp in our logs comes from the very end of log
processing right before sending it to the writer thread which executes
application callback. To check if we do not have any locking inside
logging machinery on call site let's add two timestamps to our logs. One
directly from the call site and second in formatter. Third is already
there and comes from writter as explained above.

This is a temporary commit meant to be reverted once we have results.