Update regex crate which started to fail on aarch64 windows
Suddenly regex crate started to fail building the wheel
of it by producing a ton of linker errors.

For this I've updated to latest regex which
didn't have the linker error but on aarch64
failed to invoke `build_libtelio.py` arguing
that not enough parameters are passed indicating
and issue in the package as on x86 it's fine. I bumped
it down one versio and it seems to work on both
archs.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>