Track slow nslookup performance
`nslookup` is expected to run instantly but for some
reason we observe 1-3second delay between the first
DNS packet emitted after the `nslookup` is executed.

Maybe there's a latency in `nslookup` command but another guess
is that the VM freezes/hiccups. Stracing with timestamps
would potentially show if it happens when `nslookup` is
executed.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>