Improve socket rebinding on Apple platforms
When includeAllNetworks parameter is set, and the user changes the
network interface (e.g. connects/disconnects Ethernet cable to a mac)
we end up in a no-net situation. When includeAllNetworks flag is set
we cannot call the bind method too soon, as the system will block the
traffic. The solution to this problem is to call bind method as soon
as the new network interface is fully reconfigured.