Demonize teliod
Properly daemonize the teliod daemon process.
Adds a -n flag to don't detach the daemon from the terminal.

Daemonizing changes the root of the daemon process to the daemon working
directory. This makes the relative paths become relative to the working
directory.

The daemon working directory is changed to `/var/run/teliod` or
`/run/teliod` to better organize all the files.