Pause the network monitoring while the tun interface is changed
When the set_tun is called, it results in set_iface call on NepTUN.
NepTUN internaly calls `close` on the file descriptors associated with
the old TUN interface. If the network interface starts at the wrong
moment it will call getifaddrs for the old interface with no associated
descriptors. This results in abort of the process due to an internal
assertion in glibc. This issue was found in the nat-lab tests.

To avoid this, we will pause the network monitoring for the duration of
the change of interface.