Introduce failure feedback in QNAP UI
Upon receiving POST request, after toggling
the daemon, now the result of it is transferred
back to the page so the user would get the feedback.

It's not very granular at the moment but at least it
is explicit about telling that something has failed
instead of just reseting the toggle.

This also covers the case where incorrect token is being used.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>