Fix token being hidden in teliod
Teliod uses Hidden<String> for the token
however in one case it was not dereferenced
that resulted in literal "****" to be embedded
instead of the token in core_api.rs when forming
the authorization header.

It at the moment core_api.rs doesn't contain Hidden<String>
anymore since that would be a more substantial change so for now
this file is left accepting plain &str as tokens.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>