Don't embed access token in HTML
QNAP UI(HTML), at the moment, embeds the
access token in the form. Later on when the form
is submitted, the backend handler saves whatever
value that was in a config file.

This is not a very good security model since
there's no reason for the token to be sent back to
the client in any case.

A fix for this is to explicitly allow an empty
token value which is then ignored by both the
client side validation and server side for updating.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>