Pass errors from qnap endpoint
QNAP endpoint silently swallows errors which
makes first time user, who has not yet added
a token to receive a generic "check logs" error.

A proper error is now returned back in addition
to others already existing errors.

This commit makes it so that if there's no
authentication token, upon submission
the error about it not being present is gonna be returned.

Once token is already present the allowed
empty value will be ignored and the current one will
stay.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>