Don't send back auth token to the client
authentication_token being sent back to the client
is a security risk and is totally unecessary. For this
the endpoint and QNAP UI are reworked to allow empty
authentication_token values that are ignored.

As a side effect this means there's no way to delete
the token as well but for now it's a good enough
tradeoff.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>