Make it easier to copy meshnet peer IP in QNAP
QNAP list refreshes every 2seconds which means that if
the user is in the process of copying, the focus is
lost once that happens. This makes for a very bad
copy-ip experience.

HTMX doesn't support any smart HTML replacement
feature so the whole HTML element is replaced.

A better fix would be to return JSON with data
and render everything but this is a substantial change.

As a compromise, the timer is increased from 2s->10s.
This makes meshnet peer state to be 10s late at worst
as a compromise.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>