Save CGI user intent to start the meshnet to a file
"NordSecurityMeshnet.sh start" is called
on startup of the QNAP. When restarting
QNAP "NordSecurityMeshnet.sh stop" is called.

NordSecurityMeshnet consists of two parts:
- CGI which is handled automatically for us
by the virtue of having a teliod.cgi file.
Requests are then forwarded to "teliod cgi"
command.
- Another part is Meshnet which user toggles
himself manually.

At the moment there's no indicator that the user
has intended to toggle meshnet on or off so at
startup(after restart) we cannot differentiate
at the while in NordSecurityMeshnet.sh.

For this reason the intent file is created in the app
directory when handling CGI query to toggle the meshnet
which means that user intended to either toggle
Meshnet ON or OFF. Based on that NordSecurityMeshnet.sh
is able to differentiate and either proceed fully
when "start" command is issued or short-circuit.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>