Use separate table for iproute VPN routing
Teliod with VPN and the iproute interface configuration option would
previously add routing rules to the main routing table. It worked when
testing locally but didn't work in natlab. Now the rules are written to
a separate routing table, which also aligns with the NordVPN Linux app,
and it now works in natlab. As such, the natlab test for teliod+VPN
now includes a test for using the iproute config option rather than
just the manual option