Accept LIBTELIO_VERSION env variable for version_tag
Libtelio build process assumes build promotion.

However, for teliod, the build process might be different,
for example OpenWRT uses it's own SDK and then produces a
package file and it might not be very easy to tap into it
to replace the intermediate binary before packing.

As a cheap solution, in order to support version placement
in the binary for such cases, an environment option is read
and if found then used.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>