teliod: Drop meshnet support
Meshnet is going to be sunsetted and the related API endpoints will
be unavailable starting December 2025, teliod functionality for both
meshnet and vpn connections are dependent on those endpoints.

This commit removes all the meshnet related code, refactoring the
crate to the sole purpose of serving exit node connections (vpn).

Device identity is deprecated and in its place the user service
credentials should be used to establish connection with the
vpn server. These credentials are requested to the core API.

The ip address on the status report is now fetched directly from
the interface using the selected configuration provider.

On the natlab side, core-api needed to generate and keep track of the
user service credentials for each test.  VPN containers were
only masquerading packets incoming from meshnet addresses, so
this commit adds a rule to also masquerade the non-meshnet vpn server
address.