Fix a crash when https is attempted with the default crypto provider
When derp connects to httpS it uses the default crypto provider. If
there is only one configured at build time (either ring or aws-lc-rs)
this works fine. But if both are present at compile time, rustls will
refuse to work without explicitly configuring the one which should be
the default.