Randomize server selection in teliod
The list of recommended servers that we receive from the core api is
sorted by the current load so that's why it makes sense to take the
servers from the beginning of that list.  However, the list is not
refreshed very often so it can lead to overloading the first server from
the list, especially in busy locations. To avoid it, it is recommended
to take 20 (default value) server recommendations and pick one at
random.