Prevent OpenWRT DNS leaks
By default OpenWRT receives it's settings through WAN
port via DHCP. It receives IP, gateway, subnet mask and DNS
servers to use.

It can be expected that the DNS server received from ISP will
point to the ISP itself which is also a gateway and thus will be
routed through the WAN port instead of the tunnel.

This commit allows overriding DNS servers used by dnsmasq
and it works nicely since NordSecurity has it's own DNS
servers that we can hardcode. This works naturally and DNS
requests are routed(both internal and forwareded) to the tunnel.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>