fix: select loop race condition
When client connections are accepted inside of `try_recv_quit()`, there is a possibility the connection might break,
before the reply is sent to the client and the connection is properly closed.
This can happen if the `try_recv_quit` future is canceled halfway in, after `request_nordlynx_key()` is done.

- `IsAlive` command now returns true in all stages, indicating that the daemon is already alive.
- New command response added, `DaemonInitializing`, indicating that the daemon is not ready to
process events (is fetching credentials or TelioTask has not started yet).
- Separate `accept_client_connection()` to match quickly on a `tokio::select!` branch.
The reply is sent inside the select branch, where there is no danger of it being cancelled.
- Refactored `handle_client_connection()` to handle both cases, early quit and normal processing.