wg: Trigger link detection on network changes (wireguardNT)
WireGuardNT tx counters are not increased when the interface is
disabled, which could lead to incorrect link state detection.

Add network change monitoring to link detection on Windows to improve
reliability when WireGuardNT interface state changes. When the OS
notifies about that it artificially triggers the link
detection countdown by setting first_tx_after_rx timestamp.

The change is Windows-specific and on systems without enhanced
detection it might indicate a down link state incorrectly. This
can happen if after the OS notifies about a network interface change
there isn't any rx packet for the delay amount (3s) which if the
peer is alive it shouldn't happen. Nevertheless, if it indeed happens
it is immediately fixed as soon as a packet arrives.

For further details see LLT-5073.