Remove timeouts from pq_vpn_handshake_after_nonet
Sometimes the testcase timeouts. The logs
reveal this sequence:

1) 2025-10-08 20:28:22,701 | DEBUG [DOCKER_CONE_CLIENT_1]
   Executing iptables -t filter -D OUTPUT
   --destination 10.0.100.51 -p udp -j REJECT
   --reject-with icmp-host-unreachable
2) 2025-10-08 20:28:32,715 | INFO [alpha] Test cleanup:
   Stopping tcpdump and collecting core dumps
3) 2025-10-08 20:28:32,971 | DEBUG [DOCKER_CONE_GW_1]
   Conntracker[85] reported event:     [NEW] udp
   17 30 src=192.168.101.104 dst=10.0.100.51 sport=42733
   dport=1023 [UNREPLIED] src=10.0.100.51
   dst=10.0.254.1 sport=1023 dport=42733
4) 2025-10-08 20:28:32,972 | DEBUG [DOCKER_CONE_GW_1]
   Conntracker[85] reported event:  [UPDATE] udp
   17 30 src=192.168.101.104 dst=10.0.100.51 sport=42733
   dport=1023 src=10.0.100.51 dst=10.0.254.1 sport=1023
   dport=42733

The wait for connected state happens with 10seconds and
immediatelly after dropping the iptables rule. Yet
we see that gateway reported that connection was established
at the time of cleanup which means we should have waited
just a little bit more.

We could have waited a bit more but CI runners are notorious
for sometimes being laggy so removing the timeouts is
the preferred solution for this.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>