Allow configuring DNS forwarder via OpenWRT config
OpenWRT, by default, uses DNS servers it gets from
DHCP advertisement that came from WAN port(ISP).

The DNS server advertised can be the gateway itself
which makes it so that internal and forwarded DNS
requests will be routed to the WAN port directly resulting
in a DNS leak.

Allow configuring DNS server(s) via config. Additional
property "dns" is added to the root config which accepts a
list of IPs. If the list is not provided then
NordSec VPNs are used by default(leak safe).

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>