Offset natlab startup in shards
This is possibly a temporary solution. We observe that starting docker
compose takes a lot of resources. Because our pipeline starts
9 parallel nat-lab jobs, it sometimes happen that when the job from
a previous pipeline is executed on the runner, multiple new jobs start
at the same time, all starting up docker containers. During this
startup phase the old job is experiencing timeouts We can try to offset
this startup time of the parallel runs to check if it helps.