Rotate listen-port of WG-NT adapter if server-side handshake fails
In a lengthy investigation on ocassional loss of connection we observed
that for some reason sometimes the client enters the state where it is
sending WireGuard packets towards the server, but they are not received
by the server's kernel module. Likely being dropped somewhere by
middle-hops. At the same time we observed that packets going to
different server at the time of the issue - does actually reach them,
and suspicion is that packets even to the same server, but coming from
the different source port will reach the server as well. As it will look
like "new" UDP flow to all the middle box'es, their conntrack's and so
on.

This commit adds an experimental functionality, which hooks onto WG-NT
driver logs and whenever there are at least two server-side handshakes
in a rather quick succession (10 seconds) it triggers client-side source
port change.

If such a change will recover the connectivity (expected timeout for
connectivity recovery is roughly 20-30s) then it kind of confirms the
suspicions about UDP packet flow getting corrupted somehwere down the
middle (we have captures on the client-side showing, that packets are
actually leaving the device).