Wait for ping to take effect in natlab
Link detection testcases rely on ping(icmp)
to trigger the link detection. However we often
observe that the pings are delayed for one
reason or another and usually on Windows.

The idea is instead to wait for the log line to appear instead
of relying that ping has actually succeeded.

Alternatively, it would be possible to use
ping utility itself to ensure that the ping was
actually emitted but it is more difficult to ensure.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>