NatLab: wait for ping to trigger link detection
Link detection testcases rely on ping(icmp)
to trigger the link detection. However we often
observe that the pings are delayed for one
reason or another and usually on Windows.

The idea is to wait for the log line to appear,
that way we know for sure that link detection was
triggered and thus can continue with time
sensitive assumptions.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>