Fix UAPI error tracker to be backwards compatible
Recent UAPI error tracker implementation removed the check if
interface is present when considering critical error
event.

This makes the behavior not backwards compatible with the
implementation which was present for the last 2 years.

Without the check there's possibility to emit critical
events with fewer conditions and can have real life
consequences, so better play it safe.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>